requirements.txt
# Core FastAPI and async
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# AI/ML dependencies
openai==1.3.0
langchain==0.0.350
langgraph==0.0.20
langsmith==0.0.70
langchain-google-genai
google-generativeai

# Async and utilities
aiofiles==23.2.1
python-dotenv==1.0.0

# Database integration
redis==5.0.1
sqlalchemy==2.0.23
